W5D2/Q1/requirements.txt
fastapi==0.111.0
numpy==1.26.4
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
redis==5.0.4
chromadb==0.4.24
SQLAlchemy==2.0.30
openai==1.30.1
python-dotenv==1.0.1
tenacity==8.2.3
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0 
httpx==0.26.0 
pytest==8.2.0 
jinja2==3.1.3 